Fix conflicts after rebase onto main
